dev: stop setup_test_env() nuking $CROWBAR_TEST_DIR every time
It's incredibly tedious having to wait for bundler to reinstall the
gems each time you invoke './dev tests setup'.  This happens because
the gems are installed under $CROWBAR_TEST_DIR, which setup_test_env()
nukes every time.  However, './dev tests clear' already exists to give
the user the choice to nuke it, so automatically nuking it every time
defeats the point of having that choice.